desktop: fix our AppStream file for appstreamcli validation.
Adding a TODO as date was passing through with appstream-util in relax
validation mode, but there is no such mode in appstreamcli.

Anyway since I have manual check steps for the AppStream files in
devel-docs/release-howto.txt, I never missed setting the date right (so
far! Knocks on wood…) in the AppStream metadata. Therefore let's drop
this trick for now and see how it goes.

It should fix the appstream unit check, but only when using appstreamcli
0.15.3 or over (this version is already on Debian testing).

See issue #8140.