desktop: we might not be valid project_group GNOME?
appstreamcli returns an info message:

> I: org.gimp.GIMP:3: cid-missing-affiliation-gnome org.gimp.GIMP

Checking the message, it happens because the application ID does not
start with "org.gnome." (but "org.gimp.").

Double-checking AppStream docs, a note says:

> You should only identify with an umbrella project if you use all their
> infrastructure and policies, for instance string freezes dates,
> bugtracker and source control instance.

GNOME Foundation is indeed an umbrella for us for the infrastructure and
obviously part of its policies. But we are still independant and don't
depend on GNOME's freeze dates, or release dates, or design rules
(though we try to follow what makes sense), and so on. So I guess it
means we should not set this metadata tag? 🤷 Let's drop it.

See: https://freedesktop.org/software/appstream/docs/chap-Metadata.html#tag-project_group