desktop: add (bogus) date for 2.99.12 release.
appstreamcli on CI fails with error:

> E: org.gimp.GIMP:394: release-time-missing date

I didn't have such error on my local build, because it is a new check
only recently added. The CI appstreamcli errored-out this way.

It forces us to add a bogus release date since we don't know it yet. I
chose May 29, because we usually release on Sundays, and it might soon
be time to make a dev release (so who knows, we might get lucky and it
might happen this day), though it should not be taken as a promise or a
real plan just now. I just needed to fill something in, in order to fix
the CI.

I opened a report to ask that appstreamcli proposes a relaxed check mode
where we can test on the <release> tag before knowing the future date:
https://github.com/ximion/appstream/issues/398