app: allow to disable the new version check altogether through a key…
… in the gimp-release file.

This will be useful to disable the update check for the Windows Store
even though we use the same build as the installer. All it will take
will be to append the line `check-update=false` to the file
`share/gimp/2.10/gimp-release` and it will behave as though the update
check is disabled (even though it is actually built-in).

(cherry picked from commit 9ef10c8764df53207cabe5e323c50588307547da)

Conflicts fixed:
	app/dialogs/preferences-dialog.c