plug-ins: Add .ani file import/export
This patch ports a 2.4-based plug-in from James Huang into file-ico.
It largely reads/writes the .ani header, then uses the existing
ico/cur functions to load/export the icon data.
A variable to note current file offsets are added to existing code
for this purpose.
Metadata on the cursor and author's name as well as the animation speed
are also loaded and optionally exported.